Update ruleset for iso standards (#30)

+++ src/ruleset/color_determine.h
@@ -14,7 +14,7 @@
 
 namespace color_determine {
 
-int primary_secondary_mix(const std::string& flash);
+uint16_t primary_secondary_mix(const std::string& flash);
 char missing_color(const std::string& flash);
 const bool is_primary(const char& color);
 const bool is_secondary(const char& color);
