+++ src/ruleset/inital_calculations.cc
@@ -9,7 +9,6 @@
 #include "inital_calculations.h"
 
 #include <cassert>
-#include <cstdint>
 #include <vector>
 
 #include "../math/base_36_type.h"
@@ -25,7 +24,7 @@ constexpr uint16_t base_36_factor = 36;
 constexpr uint16_t character_add = 9;
 
 int alpha(const base_36_type& serial_base) {
-    static_mod_type<int> first(mod_limit), second(mod_limit);
+    static_mod_type<int16_t> first(mod_limit), second(mod_limit);
     const auto& serial_is_number = serial_base.is_number();
     const auto& serial_number_base = serial_base.number_base();
 
@@ -41,11 +40,11 @@ int alpha(const base_36_type& serial_base) {
         second += serial_number_base[3] + character_add;
     }
 
-    return int(first + second);
+    return int16_t(first + second);
 }
 
 int bravo(const base_36_type& serial_base) {
-    static_mod_type<int> first(mod_limit);
+    static_mod_type<int16_t> first(mod_limit);
     const auto& serial_is_number = serial_base.is_number();
     const auto& serial_number_base = serial_base.number_base();
 
@@ -57,11 +56,11 @@ int bravo(const base_36_type& serial_base) {
 
     assert(serial_is_number[5]);
 
-    return int(first + serial_number_base[5]);
+    return int16_t(first + serial_number_base[5]);
 }
 
 int charlie(const base_36_type& serial_base) {
-    static_mod_type<int> first(mod_limit), second(mod_limit);
+    static_mod_type<int16_t> first(mod_limit), second(mod_limit);
     const auto& serial_is_number = serial_base.is_number();
     const auto& serial_number_base = serial_base.number_base();
 
@@ -77,11 +76,11 @@ int charlie(const base_36_type& serial_base) {
         second += serial_number_base[1] + character_add;
     }
 
-    return int(first + second);
+    return int16_t(first + second);
 }
 
 int delta(const base_36_type& serial_base) {
-    static_mod_type<int> delta(mod_limit);
+    static_mod_type<int16_t> delta(mod_limit);
     const auto& serial_is_number = serial_base.is_number();
     const auto& serial_number_base = serial_base.number_base();
 
@@ -93,7 +92,7 @@ int delta(const base_36_type& serial_base) {
         }
     }
 
-    return int(delta);
+    return int16_t(delta);
 }
 
 }  // namespace inital_calculations
